# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Crowi API
  description: API for Crowi - Markdown-based Wiki Application
  version: 2.0.0
servers:
  - url: http://localhost:3000/api/v2
    description: Local development server
  - url: https://your-crowi-instance.com/api/v2
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
...
paths:
  /app/info:
    get:
      tags:
        - app
      summary: Get public application info (site title etc.)
      responses:
        '200':
          description: Public application info
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type:
                      - string
                      - 'null'
                required:
                  - title
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - INTERNAL_ERROR
                      message:
                        type: string
                        enum:
                          - Internal server error
                    required:
                      - code
                      - message
                required:
                  - error
...